+++ packages/drivers/src/ssd1306.ts
@@ -1,8 +1,7 @@
-import { Image, Display } from "@devicescript/graphics"
+import { Image, Display, Palette } from "@devicescript/graphics"
 import { I2CDriver } from "./driver"
 import { I2CDriverOptions } from "./driver"
 import { isSimulator } from "@devicescript/core"
-import { Palette } from "@devicescript/runtime"
 
 // inspired by https://github.com/adafruit/Adafruit_CircuitPython_SSD1306/blob/main/adafruit_ssd1306.py
