+++ packages/drivers/src/st7735.ts
@@ -11,9 +11,9 @@ import {
     Image,
     SpiImageFlags,
     spiSendImage,
+    Palette
 } from "@devicescript/graphics"
 import "@devicescript/gpio"
-import { Palette } from "@devicescript/runtime"
 
 const ST7735_NOP = 0x00
 const ST7735_SWRESET = 0x01
@@ -100,7 +100,7 @@ export interface STLikeDisplayOptions extends FourWireOptions {
 }
 
 export class FourWireDriver<OPT extends FourWireOptions> {
-    constructor(public options: OPT) {}
+    constructor(public options: OPT) { }
 
     protected async sendSeq(seq: Buffer) {
         let i = 0
@@ -158,8 +158,7 @@ export class FourWireDriver<OPT extends FourWireOptions> {
 
 export class STLikeDisplayDriver
     extends FourWireDriver<STLikeDisplayOptions>
-    implements Display
-{
+    implements Display {
     public readonly palette: Palette
     private rot = 0
 
@@ -244,7 +243,7 @@ export class STLikeDisplayDriver
     }
 }
 
-export interface ST7735Options extends STLikeDisplayOptions {}
+export interface ST7735Options extends STLikeDisplayOptions { }
 
 export class ST7735Driver extends STLikeDisplayDriver {
     constructor(image: Image, options: ST7735Options) {
@@ -263,7 +262,7 @@ export class ST7735Driver extends STLikeDisplayDriver {
     }
 }
 
-export interface ST7789Options extends STLikeDisplayOptions {}
+export interface ST7789Options extends STLikeDisplayOptions { }
 
 export class ST7789Driver extends STLikeDisplayDriver {
     constructor(image: Image, options: ST7789Options) {
