+++ packages/graphics/src/index.ts
@@ -4,3 +4,4 @@ export * from "./dotmatrix"
 export * from "./image_spi"
 export * from "./display"
 export * from "./canvas"
+export * from "./palette"
