+++ packages/runtime/src/index.ts
@@ -9,3 +9,4 @@ export * from "./pixelbuffer"
 export * from "./led"
 export * from "./leddisplay"
 export * from "./palette"
+export * from "./bargraph"
